+++ projects/example/src/app/pages/home/home.ts
@@ -1,10 +1,11 @@
 import { Component } from '@angular/core';
+import { RouterLink } from '@angular/router';
 import { PkIcon } from 'ngx-pk-ui';
 import { VERSION } from '@angular/core';
 
 @Component({
   selector: 'app-home',
-  imports: [PkIcon],
+  imports: [RouterLink, PkIcon],
   templateUrl: './home.html',
 })
 export class Home {